Turned off the console log for voice state changes, is clogging up the console.
Also need to think of a soultion for why the invite new command isnt working correctly. Removing these console logs might help with keep an eye on the use of the command. Another change included in this build is the help command now reflects the invite-new command issues and that it is a work in progress.